rtl.css file and blank-page.html has prepared
An additional css file which is provided for rtl layout is made. Also a free Persian font called Vazir is added to project and it causes to show the Persian text much better.

Moreover, blank-page.html translated to Persian and the ingrediants attributes are added to rtl.css file

Furthuermore some minor translations are added to other pages